Genesis commit
This initial commit adds a minimally useful version of the library. It
contains no documentation (not even a README), and only one exported
function in the form of `readtask'.

The `readtask' function is very simple. It takes an input stream and a
`task' struct, and fills the struct with information regarding the
given task. The function returns `EOK' on success and an `errno' error
on failure.