Introduce the `writetask' function
The `writetask' function can be thought of as the inverse of `readtask'.
Instead of tasking a `task' struct and filling it with data from a
`FILE *' stream, it reads data from the struct and writes it to the
stream.

This commit also replaces the old `MAX' macro with the `max' function.
This was done because of the following code:

   mx = max(mx, strlen(*p));

If the `MAX' macro was used here, this would evaluate into the
following:

   mx = ((mx) > (strlen(*p)) ? (mx) : (strlen(*p)));

This is totally valid code, but it calls `strlen(*p)' twice which is
obviously not ideal.  Using a function solved this issue.