Error check invalid task structs
The `taskwrite' function now checks to see if both the `start' and `end'
time structs are 0'ed. In the case that they are, the function returns
EINVAL because we cannot deduce if it is either:

 - After 00:00 1900-00-00
 - Until 00:00 1900-00-00
 - On 00:00 1900-00-00

Additionally, in the case of a `From XX to YY', we test to see if `YY'
is infact after `XX'. If it isn't, then we should return EINVAL once
more.