Ensure that `XX < YY' in `From XX to YY'
This commit adds the `timecmp' function and adds checks to the
`taskread' and `taskwrite' functions to make sure that in timestamps of
the form `From XX to YY', `XX' does indeed come before `YY'.  In cases
that it doesn't, both functions will return an error value.  The
`taskwrite' function will return `EINVAL' and the `taskread' function
will return `EBADMSG'.