Don't validate dates
Previously a task was deemed invalid if both the start and end
timestamps were the epoch with the `tm_mday' field as 0, as those are
also the 0 values for the `tm' struct.  With this commit, writing those
tasks works just fine.  The rationate for this is that if the library
will do any sort of date validation, it can be done a lot better than
this.  At the moment date validation sucks, so just leave it all up to
the user.