Rename the `fp' and `tsk' to `stream' and `task'
This commit renames `fp' and `tsk' to `stream' and `task' in the
`taskread' and `taskwrite' functions. This is done in order to remain
consistent with the documenting manual pages.  The reason that these
names are being used is because as users will be using these functions,
it is important that the parameter names (which show up in
documentation, LSP, etc.) are not stupidly verbose, but are also still
readable.  A new user of the library may not know what `tsk' stands for.

A future commit will change these names in the entire codebase for
consistency.