Allocate task body buffer with `calloc'
The previous usage of `malloc' was error prone as it did not zero the
allocated buffer.  This is a problem because we write to the buffer with
`strcat', which starts writing new data from the first NUL byte.  This
caused issues when the allocated buffer returned from `malloc' didn't
return a zero'd buffer but instead a buffer with garbage data.