Properly NUL-terminate author names
As per the specification, whitespace must be squashed in authors names.
This means that "Thomas  Voss" should become "Thomas Voss".  The
`parseauthor' function does this in place, but it wasn't inserting a NUL
byte at the end of the now shortened author name, causing author names
to contain trailing characters.